mydrf/requirement.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.1
djangorestframework==3.8.2
httpie==0.9.9
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
pytz==2018.5
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
